+++ e2e/nested-frame-navigation/app/app.ts
@@ -1,0 +1,5 @@
+import * as application from "tns-core-modules/application";
+
+application.run({ moduleName: "app-root" });
+// application.run({ moduleName: "tab-root" });
+// application.run({ moduleName: "layout-root" });