+++ e2e/nested-frame-navigation/app/shared/item.ts
@@ -1,0 +1,6 @@
+export interface Item {
+    id: number;
+    name: string;
+    description: string;
+    transition: string;
+}