+++ tests/app/ui/helper.ts
@@ -150,9 +150,13 @@ export function buildUIWithWeakRefAndInteract<T extends View>(createFunc: () => 
         // Could cause GC on the next call.
         // NOTE: Don't replace this with forceGC();
         new ArrayBuffer(4 * 1024 * 1024);
+
+        // An additional GC and wait are needed since WebKit upgrade to version 12.0
+        // (TEXT-FIELD.testMemoryLeak test started failing sporadically)
+        utils.GC();
+        TKUnit.wait(0.1);
     }
     utils.GC();
-
     try {
         TKUnit.assert(!weakRef.get(), weakRef.get() + " leaked!");
         done(null);
@@ -209,8 +213,8 @@ export function waitUntilNavigatedTo(page: Page, action: Function) {
     TKUnit.waitUntilReady(() => completed, 5);
 }
 
-export function waitUntilNavigatedFrom(action: Function) {
-    const currentPage = frame.topmost().currentPage;
+export function waitUntilNavigatedFrom(action: Function, topFrame?: frame.Frame) {
+    const currentPage = topFrame ? topFrame.currentPage : frame.topmost().currentPage;
     let completed = false;
     function navigatedFrom(args) {
         args.object.page.off("navigatedFrom", navigatedFrom);
@@ -226,19 +230,19 @@ export function waitUntilLayoutReady(view: View): void {
     TKUnit.waitUntilReady(() => view.isLayoutValid);
 }
 
-export function navigateWithEntry(entry: frame.NavigationEntry): Page {
+export function navigateWithEntry(entry: frame.NavigationEntry, topFrame?: frame.Frame): Page {
     const page = createViewFromEntry(entry) as Page;
     entry.moduleName = null;
     entry.create = function () {
         return page;
     };
 
-    waitUntilNavigatedFrom(() => frame.topmost().navigate(entry));
+    waitUntilNavigatedFrom(() => topFrame ? topFrame.navigate(entry) : frame.topmost().navigate(entry));
     return page;
 }
 
-export function goBack() {
-    waitUntilNavigatedFrom(() => frame.topmost().goBack());
+export function goBack(topFrame?: frame.Frame) {
+    waitUntilNavigatedFrom(() => topFrame ? topFrame.goBack() : frame.topmost().goBack());
 }
 
 export function assertAreClose(actual: number, expected: number, message: string): void {
@@ -683,13 +687,13 @@ function setupSetters(): void {
     cssSetters.set("selectedTabTextColor", "red");
     cssSetters.set("androidSelectedTabHighlightColor", "red");
 
-    // ListView-specific props 
+    // ListView-specific props
     cssSetters.set("separatorColor", "red");
 
     // SegmentedBar-specific props
     cssSetters.set("selectedBackgroundColor", "red");
 
-    // Page-specific props 
+    // Page-specific props
     cssSetters.set("statusBarStyle", "light");
     cssSetters.set("androidStatusBarBackground", "red");