+++ tns-core-modules/ui/tab-view/tab-view.android.ts
@@ -45,7 +45,7 @@ function initializeNativeClasses() {
         return;
     }
 
-    class TabFragmentImplementation extends android.support.v4.app.Fragment {
+    class TabFragmentImplementation extends org.nativescript.widgets.FragmentBase {
         private tab: TabView;
         private index: number;
 
@@ -55,7 +55,6 @@ function initializeNativeClasses() {
         }
 
         static newInstance(tabId: number, index: number): TabFragmentImplementation {
-
             const args = new android.os.Bundle();
             args.putInt(TABID, tabId);
             args.putInt(INDEX, index);
@@ -78,10 +77,6 @@ function initializeNativeClasses() {
             const tabItem = this.tab.items[this.index];
 
             return tabItem.view.nativeViewProtected;
-        }
-
-        public onDestroyView() {
-            super.onDestroyView();
         }
     }
 
@@ -200,10 +195,7 @@ function initializeNativeClasses() {
         }
 
         finishUpdate(container: android.view.ViewGroup): void {
-            if (this.mCurTransaction != null) {
-                (<any>this.mCurTransaction).commitNowAllowingStateLoss();
-                this.mCurTransaction = null;
-            }
+            this._commitCurrentTransaction();
         }
 
         isViewFromObject(view: android.view.View, object: java.lang.Object): boolean {
@@ -211,6 +203,9 @@ function initializeNativeClasses() {
         }
 
         saveState(): android.os.Parcelable {
+            // Commit the current transaction on save to prevent "No view found for id 0xa" exception on restore.
+            // Related to: https://github.com/NativeScript/NativeScript/issues/6466
+            this._commitCurrentTransaction();
             return null;
         }
 
@@ -221,8 +216,15 @@ function initializeNativeClasses() {
         getItemId(position: number): number {
             return position;
         }
-    }
 
+        private _commitCurrentTransaction() {
+            if (this.mCurTransaction != null) {
+                this.mCurTransaction.commitNowAllowingStateLoss();
+                this.mCurTransaction = null;
+            }
+        }
+    }
+    
     PagerAdapter = FragmentPagerAdapter;
 }
 
@@ -430,6 +432,10 @@ export class TabView extends TabViewBase {
             nativeView.addRow(new org.nativescript.widgets.ItemSpec(1, org.nativescript.widgets.GridUnitType.star));
 
             viewPager.setLayoutParams(lp);
+
+            if (!this.androidSwipeEnabled) {
+                viewPager.setSwipePageEnabled(false);
+            }
         } else {
             nativeView.addRow(new org.nativescript.widgets.ItemSpec(1, org.nativescript.widgets.GridUnitType.star));
             nativeView.addRow(new org.nativescript.widgets.ItemSpec(1, org.nativescript.widgets.GridUnitType.auto));
@@ -482,7 +488,7 @@ export class TabView extends TabViewBase {
     public _loadUnloadTabItems(newIndex: number) {
         const items = this.items;
         const lastIndex = this.items.length - 1;
-        const offsideItems = this.androidTabsPosition === "top" ? this.androidOffscreenTabLimit : 0;
+        const offsideItems = this.androidTabsPosition === "top" ? this.androidOffscreenTabLimit : 1;
 
         let toUnload = [];
         let toLoad = [];
@@ -506,7 +512,7 @@ export class TabView extends TabViewBase {
         const newItem = items[newIndex];
         const selectedView = newItem && newItem.view;
         if (selectedView instanceof Frame) {
-            selectedView._pushInFrameStack();
+            selectedView._pushInFrameStackRecursive();
         }
 
         toLoad.forEach(index => {
@@ -549,8 +555,13 @@ export class TabView extends TabViewBase {
     }
 
     public _onRootViewReset(): void {
-        this.disposeCurrentFragments();
         super._onRootViewReset();
+        
+        // call this AFTER the super call to ensure descendants apply their rootview-reset logic first
+        // i.e. in a scenario with tab frames let the frames cleanup their fragments first, and then
+        // cleanup the tab fragments to avoid
+        // android.content.res.Resources$NotFoundException: Unable to find resource ID #0xfffffff6
+        this.disposeCurrentFragments();
     }
 
     private disposeCurrentFragments(): void {